refactor: mk Kafka metrics pkg public
