+++ pkg/kafka/metrics/stats/testdata/stats.json
